+++ changelog.md
@@ -9,6 +9,7 @@ Codion Change Log
 ### is.codion.framework.domain
 - EntityDefinition.Builder.descriptionResource() added.
 - AttributeDefinition.Builder.descriptionResource() added.
+- DefaultEntityBuilder bug fixed, now initializes primary key column values to null for entities with non-generated primary keys, since the default Entity.exists() predicate relies on the original primary key being null.
 ### is.codion.framework.domain.db
 - SchemaDomain.SchemaSettings.lowerCaseIdentifiers() added.
 ### is.codion.plugin.swing.robot